created unit and integration tests for task service

+++ src/main/java/com/qa/todo/service/TaskService.java
@@ -56,6 +56,9 @@ public class TaskService {
 	
 	//delete
 	public boolean delete(Long id) {
+		if(!this.repo.existsById(id)) {
+			throw new TaskNotFoundException();
+		}
 		this.repo.deleteById(id);
 		return !this.repo.existsById(id);
 	}
